backend/requirements.txt
Flask==3.1.0
flask-cors==6.0.0
PyJWT==2.9.0
pymongo==3.11.4
bcrypt==4.3.0
python-dotenv==1.0.0
pydantic
google-auth 
google-auth-oauthlib
google-auth-httplib2
python-dateutil
google-api-python-client
dnspython==2.7.0
Pillow==11.3.0
flask-socketio==5.3.6
python-socketio==5.8.0
Flask-Mail==0.10.0
itsdangerous==2.2.0
